Change TypeId size to 128 bits
https://github.com/rust-lang/rust/pull/109953 made std TypeId 128 bits
and its new Hash impl prevents the use of TypeIdHasher
so I'll be using a pointer cast going forward

+++ src/scheduler/into_workload.rs
@@ -169,7 +169,7 @@ where
             let unique_id = unique_id();
 
             let name = Box::new(WorkloadLabel {
-                type_id: TypeId(unique_id),
+                type_id: TypeId(unique_id as u128),
                 name: unique_id.to_string().as_label(),
             });
 
@@ -211,7 +211,7 @@ macro_rules! impl_into_workload {
                 let unique_id = unique_id();
 
                 let name = Box::new(WorkloadLabel {
-                    type_id: TypeId(unique_id),
+                    type_id: TypeId(unique_id as u128),
                     name: unique_id.to_string().as_label(),
                 });
 
@@ -241,7 +241,7 @@ macro_rules! impl_into_workload {
                 let unique_id = unique_id();
 
                 let name = Box::new(WorkloadLabel {
-                    type_id: TypeId(unique_id),
+                    type_id: TypeId(unique_id as u128),
                     name: unique_id.to_string().as_label(),
                 });
